Bump version to 2.0.0 and refactor Mattermost client
Updated `GitVersion.yml` to version 2.0.0.
Refactored `MattermostClient` into partial classes for better organization.
Removed obsolete methods from `IMattermostClient` and added new methods for channel operations.
Enhanced tests in `MattermostClientTests.cs` to verify channel information retrieval.
Updated `Team` model to public and added new properties.
Introduced methods for user information retrieval in `MattermostClient.Users.cs`.
Overall, these changes improve structure, add features, and remove deprecated methods.
