+++ GitVersion.yml
@@ -1,1 +1,1 @@
-next-version: 1.0.0
+next-version: 2.0.0
