+++ Sources/Mattermost/MattermostClient.cs
@@ -1,26 +1,17 @@
 ﻿using System;
 using System.IO;
 using System.Linq;
-using System.Text;
 using System.Net.Http;
 using System.Threading;
-using System.Text.Json;
 using Mattermost.Enums;
 using Mattermost.Events;
-using Mattermost.Models;
-using Mattermost.Helpers;
 using Mattermost.Constants;
 using Mattermost.Extensions;
 using Mattermost.Exceptions;
 using System.Net.WebSockets;
 using System.Threading.Tasks;
 using System.Net.Http.Headers;
-using Mattermost.Models.Posts;
-using Mattermost.Models.Teams;
 using Mattermost.Models.Users;
-using System.Collections.Generic;
-using Mattermost.Models.Channels;
-using Mattermost.Models.Responses;
 using Mattermost.Models.Responses.Websocket;
 
 namespace Mattermost
@@ -28,7 +19,7 @@ namespace Mattermost
     /// <summary>
     /// .NET API client for Mattermost servers with websocket polling.
     /// </summary>
-    public class MattermostClient : IMattermostClient, IDisposable
+    public partial class MattermostClient : IMattermostClient, IDisposable
     {
         /// <summary>
         /// Called when client is connected to server WebSocket after
@@ -262,18 +253,6 @@ namespace Mattermost
         }
 
         /// <summary>
-        /// Login with specified API key - personal access token, see <see href="https://developers.mattermost.com/integrate/reference/personal-access-token/"/>
-        /// </summary>
-        /// <param name="apiKey"> API key, ex. bot token or personal access token. </param>
-        /// <returns>Autorized <see cref="User"/> object.</returns>
-        /// <exception cref="AuthorizationException">Throws if API key is invalid or server response is not successful.</exception>
-        [Obsolete("Use constructor with API key instead, ex. new MattermostClient(serverUrl, apiKey)")]
-        public Task<User> LoginAsync(string apiKey)
-        {
-            return LoginWithApiKeyAsync(apiKey);
-        }
-
-        /// <summary>
         /// Logout from server.
         /// </summary>
         /// <returns> Task representing logout operation. </returns>
@@ -289,481 +268,6 @@ namespace Mattermost
             }
             await StopReceivingAsync();
             _http.DefaultRequestHeaders.Authorization = null;
-        }
-
-        /// <summary>
-        /// Get current authorized user information and force update of <see cref="CurrentUserInfo"/>.
-        /// </summary>
-        /// <returns> Authorized user information. </returns>
-        public async Task<User> GetMeAsync()
-        {
-            CheckDisposed();
-            CheckAuthorized();
-            var response = await _http.GetAsync(Routes.Users + "/me");
-            response.EnsureSuccessStatusCode();
-            string json = await response.Content.ReadAsStringAsync();
-            _userInfo = JsonSerializer.Deserialize<User>(json)!;
-            return _userInfo;
-        }
-
-
-        /// <summary>
-        /// Send message to specified channel identifier.
-        /// </summary>
-        /// <param name="channelId"> Channel identifier. </param>
-        /// <param name="message"> Message text (Markdown supported). </param>
-        /// <param name="replyToPostId"> Reply to post (optional) </param>
-        /// <param name="priority"> Set message priority </param>
-        /// <param name="files"> Attach files to post. </param>
-        /// <returns> Created post. </returns>
-        /// <exception cref="ArgumentOutOfRangeException">Thrown when message length exceed maximum limit of characters, see <see cref="MattermostApiLimits.MaxPostMessageLength"/>.</exception>
-        [Obsolete("Use CreatePostAsync instead.")]
-        public Task<Post> SendMessageAsync(string channelId, string message = "",
-            string replyToPostId = "", MessagePriority priority = MessagePriority.Empty,
-            IEnumerable<string>? files = null) => CreatePostAsync(channelId, message, replyToPostId, priority, files);
-
-        /// <summary>
-        /// Send message to specified channel identifier.
-        /// </summary>
-        /// <param name="channelId"> Channel identifier. </param>
-        /// <param name="message"> Message text (Markdown supported). </param>
-        /// <param name="replyToPostId"> Reply to post (optional) </param>
-        /// <param name="priority"> Set message priority </param>
-        /// <param name="files"> Attach files to post. </param>
-        /// <param name="props"> A general JSON property bag to attach to the post. </param>
-        /// <returns> Created post. </returns>
-        /// <exception cref="ArgumentOutOfRangeException">Thrown when message length exceed maximum limit of characters, see <see cref="MattermostApiLimits.MaxPostMessageLength"/>.</exception>
-        public async Task<Post> CreatePostAsync(string channelId, string message = "",
-            string replyToPostId = "", MessagePriority priority = MessagePriority.Empty,
-            IEnumerable<string>? files = null, IDictionary<string, object>? props = null)
-        {
-            if (message.Length > MattermostApiLimits.MaxPostMessageLength)
-            {
-                throw new ArgumentOutOfRangeException(nameof(message),
-                    $"The message length exceeds the maximum number of characters allowed ({message.Length} > {MattermostApiLimits.MaxPostMessageLength})");
-            }
-
-            CheckDisposed();
-            CheckAuthorized();
-            Dictionary<string, object> metadata = new Dictionary<string, object>();
-            if (priority != MessagePriority.Empty)
-            {
-                metadata.Add("priority", new
-                {
-                    priority = priority.ToString().ToLower(),
-                    requested_ack = false
-                });
-            }
-
-            var body = new
-            {
-                message,
-                channel_id = channelId,
-                root_id = replyToPostId,
-                metadata,
-                file_ids = files,
-                props
-            };
-            string json = JsonSerializer.Serialize(body);
-            StringContent content = new StringContent(json, Encoding.UTF8, "application/json");
-            var response = await _http.PostAsync(Routes.Posts, content);
-            response = response.EnsureSuccessStatusCode();
-            json = await response.Content.ReadAsStringAsync();
-            return JsonSerializer.Deserialize<Post>(json)!;
-        }
-
-        /// <summary>
-        /// Update message text for specified post identifier.
-        /// </summary>
-        /// <param name="postId"> Post identifier. </param>
-        /// <param name="newText"> New message text (Markdown supported). </param>
-        /// <param name="props"> A general JSON property bag to attach to the post. </param>
-        /// <returns> Updated post. </returns>
-        /// <exception cref="ArgumentOutOfRangeException">Thrown when message length exceed maximum limit of characters, see <see cref="MattermostApiLimits.MaxPostMessageLength"/>.</exception>
-        public async Task<Post> UpdatePostAsync(string postId, string newText, IDictionary<string, object>? props = null)
-        {
-            if (newText.Length > MattermostApiLimits.MaxPostMessageLength)
-            {
-                throw new ArgumentOutOfRangeException(nameof(newText),
-                    $"The message length exceeds the maximum number of characters allowed ({newText.Length} > {MattermostApiLimits.MaxPostMessageLength})");
-            }
-
-            CheckDisposed();
-            CheckAuthorized();
-            var body = new
-            {
-                message = newText,
-                props
-            };
-            string json = JsonSerializer.Serialize(body);
-            StringContent content = new StringContent(json, Encoding.UTF8, "application/json");
-            var response = await _http.PutAsync(Routes.Posts + "/" + postId + "/patch", content);
-            response = response.EnsureSuccessStatusCode();
-            json = await response.Content.ReadAsStringAsync();
-            return JsonSerializer.Deserialize<Post>(json)!;
-        }
-
-        /// <summary>
-        /// Delete post with specified post identifier.
-        /// </summary>
-        /// <param name="postId"> Post identifier. </param>
-        /// <returns> True if deleted, otherwise false. </returns>
-        public async Task<bool> DeletePostAsync(string postId)
-        {
-            CheckDisposed();
-            CheckAuthorized();
-            var response = await _http.DeleteAsync(Routes.Posts + "/" + postId);
-            return response.IsSuccessStatusCode;
-        }
-
-        /// <summary>
-        /// Get a page of posts in a channel.
-        /// </summary>
-        /// <param name="channelId"> Channel identifier. </param>
-        /// <param name="page"> The page to select. </param>
-        /// <param name="perPage"> The number of posts per page. </param>
-        /// <param name="beforePostId"> A post id to select the posts that came before this one. </param>
-        /// <param name="afterPostId"> A post id to select the posts that came after this one. </param>
-        /// <param name="includeDeleted"> Whether to include deleted posts or not. Must have system admin permissions. </param>
-        /// <param name="since"> Time to select modified posts after. </param>
-        /// <returns> ChannelPosts object with posts. </returns>
-        public async Task<ChannelPostsResponse> GetChannelPostsAsync(string channelId, int page = 0,
-            int perPage = 60, string? beforePostId = null, string? afterPostId = null,
-            bool includeDeleted = false, DateTime? since = null)
-        {
-            CheckDisposed();
-            CheckAuthorized();
-
-            string query = QueryHelpers.BuildChannelPostsQuery(page, perPage, beforePostId, afterPostId, includeDeleted, since);
-            string url = $"{Routes.Channels}/{channelId}/posts?{query}";
-            var response = await _http.GetAsync(url);
-            response = response.EnsureSuccessStatusCode();
-            string json = await response.Content.ReadAsStringAsync();
-            return JsonSerializer.Deserialize<ChannelPostsResponse>(json)
-                ?? throw new MattermostClientException("Failed to deserialize channel posts response");
-        }
-
-        /// <summary>
-        /// Get posts related to specified post identifier in thread format.
-        /// </summary>
-        /// <param name="postId"> Post identifier to get thread posts. </param>
-        /// <param name="fromPostId"> Post identifier to start from. </param>
-        /// <returns> Collection of posts in thread format. </returns>
-        public async Task<IEnumerable<Post>> GetThreadPostsAsync(string postId, string? fromPostId = null)
-        {
-            CheckDisposed();
-            CheckAuthorized();
-            string url = $"{Routes.Posts}/{postId}/thread";
-            if (!string.IsNullOrEmpty(fromPostId))
-            {
-                url += $"?fromPost={fromPostId}";
-            }
-            var response = await _http.GetAsync(url);
-            response = response.EnsureSuccessStatusCode();
-            string json = await response.Content.ReadAsStringAsync();
-            var result = JsonSerializer.Deserialize<ChannelPostsResponse>(json)
-                ?? throw new MattermostClientException("Failed to deserialize thread posts response");
-            return result.Posts.Values.OrderBy(p => p.CreatedAt);
-        }
-
-        /// <summary>
-        /// Create group channel with specified users.
-        /// </summary>
-        /// <param name="userIds"> Participant users. </param>
-        /// <returns> Created channel info. </returns>
-        public async Task<Channel> CreateGroupChannelAsync(params string[] userIds)
-        {
-            CheckDisposed();
-            CheckAuthorized();
-            string json = JsonSerializer.Serialize(userIds);
-            StringContent content = new StringContent(json, Encoding.UTF8, "application/json");
-            var response = await _http.PostAsync(Routes.GroupChannels, content);
-            response = response.EnsureSuccessStatusCode();
-            json = await response.Content.ReadAsStringAsync();
-            var result = JsonSerializer.Deserialize<Channel>(json)!;
-
-            //TODO: create channel link
-            var team = await GetTeamAsync(result.TeamId);
-            result.Link = ServerAddress + team.Name + "/channels/" + result.Id;
-            //
-
-            return result;
-        }
-
-        /// <summary>
-        /// Create simple channel with specified users.
-        /// </summary>
-        /// <param name="teamId"> Team identifier. </param>
-        /// <param name="name"> Channel name. </param>
-        /// <param name="displayName"> Channel display name. </param>
-        /// <param name="channelType"> Channel type: open or private. </param>
-        /// <param name="purpose"> Channel purpose (optional). </param>
-        /// <param name="header"> Channel header (optional). </param>
-        /// <returns> Created channel info. </returns>
-        public async Task<Channel> CreateChannelAsync(string teamId, string name, string displayName,
-            ChannelType channelType, string purpose = "", string header = "")
-        {
-            CheckDisposed();
-            CheckAuthorized();
-            const int maxChannelDisplayNameLength = 64;
-            if (displayName.Length > maxChannelDisplayNameLength)
-            {
-                throw new ArgumentException("Display name is too long", nameof(displayName));
-            }
-            var body = new
-            {
-                team_id = teamId,
-                name,
-                display_name = displayName,
-                purpose,
-                header,
-                type = channelType.ToChannelChar()
-            };
-            string json = JsonSerializer.Serialize(body);
-            StringContent content = new StringContent(json, Encoding.UTF8, "application/json");
-            var response = await _http.PostAsync(Routes.Channels, content);
-            response = response.EnsureSuccessStatusCode();
-            json = await response.Content.ReadAsStringAsync();
-            var result = JsonSerializer.Deserialize<Channel>(json)!;
-            var team = await GetTeamAsync(teamId);
-            result.Link = ServerAddress + team.Name + "/channels/" + result.Id;
-            return result;
-        }
-
-        private async Task<Team> GetTeamAsync(string teamId)
-        {
-            string url = Routes.Teams + "/" + teamId;
-            var response = await _http.GetAsync(url);
-            response = response.EnsureSuccessStatusCode();
-            string json = await response.Content.ReadAsStringAsync();
-            return JsonSerializer.Deserialize<Team>(json)!;
-        }
-
-        /// <summary>
-        /// Get user by identifier.
-        /// </summary>
-        /// <param name="userId"> User identifier. </param>
-        /// <returns> User information. </returns>
-        public async Task<User> GetUserAsync(string userId)
-        {
-            CheckDisposed();
-            CheckAuthorized();
-            string url = Routes.Users + "/" + userId;
-            string json = await _http.GetStringAsync(url);
-            User userInfo = JsonSerializer.Deserialize<User>(json)!;
-            return userInfo;
-        }
-
-        /// <summary>
-        /// Get user by username.
-        /// </summary>
-        /// <param name="username"> Username. </param>
-        /// <returns> User information. </returns>
-        public async Task<User> GetUserByUsernameAsync(string username)
-        {
-            CheckDisposed();
-            CheckAuthorized();
-            string url = Routes.Users + "/username/" + username.Replace("@", string.Empty).Trim();
-            string json = await _http.GetStringAsync(url);
-            User userInfo = JsonSerializer.Deserialize<User>(json)!;
-            return userInfo;
-        }
-
-        /// <summary>
-        /// Add user to channel.
-        /// </summary>
-        /// <param name="channelId"> Channel identifier. </param>
-        /// <param name="userId"> User identifier. </param>
-        /// <returns> Channel user information. </returns>
-        public async Task<ChannelUserInfo> AddUserToChannelAsync(string channelId, string userId)
-        {
-            CheckDisposed();
-            CheckAuthorized();
-            string url = Routes.Channels + $"/{channelId}/members";
-            var body = new
-            {
-                user_id = userId
-            };
-            string json = JsonSerializer.Serialize(body);
-            StringContent content = new StringContent(json, Encoding.UTF8, "application/json");
-            var response = await _http.PostAsync(url, content);
-            response = response.EnsureSuccessStatusCode();
-            json = await response.Content.ReadAsStringAsync();
-            return JsonSerializer.Deserialize<ChannelUserInfo>(json)!;
-        }
-
-        /// <summary>
-        /// Delete user from channel.
-        /// </summary>
-        /// <param name="channelId"> Channel identifier. </param>
-        /// <param name="userId"> User identifier. </param>
-        /// <returns> True if deleted, otherwise false. </returns>
-        public async Task<bool> DeleteUserFromChannelAsync(string channelId, string userId)
-        {
-            CheckDisposed();
-            CheckAuthorized();
-            string url = Routes.Channels + $"/{channelId}/members/{userId}";
-            var response = await _http.DeleteAsync(url);
-            return response.IsSuccessStatusCode;
-        }
-
-        /// <summary>
-        /// Upload new file.
-        /// </summary>
-        /// <param name="channelId"> Channel where file will be posted. </param>
-        /// <param name="filePath"> File fullname on local device. </param>
-        /// <param name="progressChanged"> Uploading progress callback in percents - from 0 to 100. </param>
-        /// <returns> Created file details. </returns>
-        public async Task<FileDetails> UploadFileAsync(string channelId, string filePath, Action<int>? progressChanged = null)
-        {
-            FileInfo fileInfo = new FileInfo(filePath);
-            using var fs = fileInfo.OpenRead();
-            return await UploadFileAsync(channelId, fileInfo.Name, fs, progressChanged);
-        }
-
-        /// <summary>
-        /// Upload new file.
-        /// </summary>
-        /// <param name="channelId"> Channel where file will be posted. </param>
-        /// <param name="fileName"> Name of the uploaded file. </param>
-        /// <param name="stream"> File content. </param>
-        /// <param name="progressChanged"> Uploading progress callback in percents - from 0 to 100. </param>
-        /// <returns> Created file details. </returns>
-        public async Task<FileDetails> UploadFileAsync(string channelId, string fileName, Stream stream, Action<int>? progressChanged = null)
-        {
-            CheckDisposed();
-            CheckAuthorized();
-            string url = $"{Routes.Files}?channel_id={channelId}";
-            MultipartFormDataContent content = new MultipartFormDataContent();
-            StreamContent file = new StreamContent(stream);
-            content.Add(file, "files", fileName);
-            CancellationTokenSource cts = new CancellationTokenSource();
-            if (progressChanged != null)
-            {
-                StartProgressTracker(stream, cts.Token, progressChanged);
-            }
-            var result = await _http.PostAsync(url, content);
-            result = result.EnsureSuccessStatusCode();
-            cts.Cancel();
-            string json = await result.Content.ReadAsStringAsync();
-            return JsonSerializer.Deserialize<FileResponse>(json)!.Files.First();
-        }
-
-        /// <summary>
-        /// Get file details by specified identifier.
-        /// </summary>
-        /// <param name="fileId"> File identifier. </param>
-        /// <returns> File details. </returns>
-        public async Task<FileDetails> GetFileDetailsAsync(string fileId)
-        {
-            CheckDisposed();
-            CheckAuthorized();
-            string url = Routes.Files + "/" + fileId + "/info";
-            string json = await _http.GetStringAsync(url);
-            return JsonSerializer.Deserialize<FileDetails>(json)!;
-        }
-
-        /// <summary>
-        /// Find channel by channel name and team name or identifier.
-        /// </summary>
-        /// <param name="teamIdOrName"> Team name or identifier where channel exists. </param>
-        /// <param name="channelName"> Channel name. </param>
-        /// <param name="isTeamId"> True if teamIdOrName is team identifier, otherwise false (team name). Default is true. </param>
-        /// <param name="includeDeleted"> Include deleted channels in search, default is true. </param>
-        /// <returns> Channel info, or null if team or channel not found. </returns>
-        public async Task<Channel?> FindChannelByNameAsync(string teamIdOrName, string channelName, bool isTeamId = true, bool includeDeleted = true)
-        {
-            CheckDisposed();
-            CheckAuthorized();
-            string escapedTeamIdOrName = Uri.EscapeDataString(teamIdOrName);
-            string escapedChannelName = Uri.EscapeDataString(channelName);
-            string url = isTeamId
-                ? Routes.Teams + $"/{escapedTeamIdOrName}/channels/name/{escapedChannelName}?include_deleted={includeDeleted}"
-                : Routes.Teams + $"/name/{escapedTeamIdOrName}/channels/name/{escapedChannelName}?include_deleted={includeDeleted}";
-            var response = await _http.GetAsync(url);
-            if (!response.IsSuccessStatusCode)
-            {
-                return null;
-            }
-            string json = await response.Content.ReadAsStringAsync();
-            return JsonSerializer.Deserialize<Channel>(json);
-        }
-
-        /// <summary>
-        /// Set call state for channel identifier.
-        /// </summary>
-        /// <param name="channelId"> Channel identifier where calls must be in specified state. </param>
-        /// <param name="isCallsEnabled"> New state. </param>
-        public async Task<bool> SetChannelCallStateAsync(string channelId, bool isCallsEnabled)
-        {
-            CheckDisposed();
-            CheckAuthorized();
-            string url = Routes.Plugins + "/com.mattermost.calls/" + channelId;
-            var body = new
-            {
-                enabled = isCallsEnabled
-            };
-            string json = JsonSerializer.Serialize(body);
-            StringContent content = new StringContent(json, Encoding.UTF8, "application/json");
-            var response = await _http.PostAsync(url, content);
-            return response.IsSuccessStatusCode;
-        }
-
-        /// <summary>
-        /// Archive channel by specified channel identifier.
-        /// </summary>
-        /// <param name="channelId"> Channel identifier. </param>
-        public async Task<bool> ArchiveChannelAsync(string channelId)
-        {
-            CheckDisposed();
-            CheckAuthorized();
-            string url = Routes.Channels + "/" + channelId;
-            var response = await _http.DeleteAsync(url);
-            return response.IsSuccessStatusCode;
-        }
-
-        /// <summary>
-        /// Get post by identifier.
-        /// </summary>
-        /// <param name="postId"> Post identifier. </param>
-        /// <returns> Post information. </returns>
-        public async Task<Post> GetPostAsync(string postId)
-        {
-            CheckDisposed();
-            CheckAuthorized();
-            string url = Routes.Posts + "/" + postId;
-            string json = await _http.GetStringAsync(url);
-            return JsonSerializer.Deserialize<Post>(json)!;
-        }
-
-        /// <summary>
-        /// Get file by identifier.
-        /// </summary>
-        /// <param name="fileId"> File identifier. </param>
-        /// <returns> File bytes. </returns>
-        public async Task<byte[]> GetFileAsync(string fileId)
-        {
-            CheckDisposed();
-            CheckAuthorized();
-            string url = Routes.Files + "/" + fileId;
-            var response = await _http.GetAsync(url);
-            response = response.EnsureSuccessStatusCode();
-            return await response.Content.ReadAsByteArrayAsync();
-        }
-
-        /// <summary>
-        /// Get file stream by identifier.
-        /// </summary>
-        /// <param name="fileId"> File identifier. </param>
-        /// <returns> File stream. </returns>
-        public async Task<Stream> GetFileStreamAsync(string fileId)
-        {
-            CheckDisposed();
-            CheckAuthorized();
-            string url = Routes.Files + "/" + fileId;
-            var response = await _http.GetAsync(url);
-            response = response.EnsureSuccessStatusCode();
-            return await response.Content.ReadAsStreamAsync();
         }
 
         /// <summary>
@@ -783,8 +287,6 @@ namespace Mattermost
                 }
             }
         }
-
-        #region Private methods
 
         private Task HandleResponseAsync(WebsocketMessage response, CancellationToken cancellationToken)
         {
@@ -969,7 +471,5 @@ namespace Mattermost
             _userInfo = result.GetResponse<User>();
             return _userInfo;
         }
-
-        #endregion
     }
 }
